Completed Audio Juice
Changed some scripts (DragItem...) to add <public AudioSource sound> and their respective lines to trigger the sfx.

Also, added MusicManager and SoundManager to Project Heirarchy.

Music Manager makes the transitions between intro music and garden music. Also, switches music on or off, when going to AR screen.

It seems everything is working. The only issue was with the AR boxes. Somehow, the sound plays first, and then the animation of the box is triggered.